Correct spec file errors.
Tito places the RPM release in the tag and GitHub uses the tag name to
create the archive tarball.  The result is an archive named
salmon-1.0.0-1 for example.  We can't actually reference the pure
release number in the spec file though since the release has a %{dist}
tag applied to it.  We could use a global macro to store the pure number
and then use that global in both the "Release" and "Source0" lines, but
then Tito wouldn't automatically increment the release number.  So for
now, we just can't use the GitHub archive URL.